+++ app/Exceptions/User/LoginFailedException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace App\Exceptions\User;
+
+use Exception;
+
+class LoginFailedException extends Exception
+{
+}
