+++ app/Http/Controllers/Api/AuthController.php
@@ -2,23 +2,23 @@
 
 namespace App\Http\Controllers\Api;
 
+use App\Exceptions\User\LoginFailedException;
+use App\Http\Resources\UserResource;
 use App\Models\User;
 use App\Http\Controllers\Controller;
-use App\Http\Requests\LoginUser;
+use App\Http\Requests\User\LoginUserRequest;
 use App\Http\Requests\RegisterUser;
+use App\UseCases\User\LoginUser;
 use App\ViewModels\UserViewModel;
 use Illuminate\Support\Facades\Auth;
 
 class AuthController extends Controller
 {
-    public function login(LoginUser $request)
+    public function login(LoginUserRequest $request, LoginUser $usecase)
     {
-        $credentials = [
-            'email' => $request->input('user.email'),
-            'password' => $request->input('user.password'),
-        ];
-
-        if (!$token = Auth::attempt($credentials)) {
+        try {
+            $usecase($request);
+        } catch (LoginFailedException $exception) {
             return response()->json([
                 'errors' => [
                     'message' => 'email or password is invalid'
@@ -26,7 +26,7 @@ class AuthController extends Controller
             ], 422);
         }
 
-        return new UserViewModel(Auth::user());
+        return new UserResource(Auth::user());
     }
 
     public function register(RegisterUser $request)
